Update and rename AboutUs.js to Courses.js
AboutUs screens updated to show course cards (item-card component)
Renamed to Courses.js
